Widen Proxmox byte and timestamp fields to Int64 for 32-bit watchOS
Apple Watch (Series 4-8/SE/Ultra1) runs the arm64_32 ABI where Swift Int is 32-bit (max ~2.1e9). Proxmox byte counts (multi-GB memory/disk, cumulative network) exceed that, so JSONDecoder threw dataCorrupted ("Number ... not representable in Swift") on real watch hardware, while iPhone (64-bit) and the watch Simulator (64-bit) decoded fine.

Widen all byte-count and epoch-timestamp fields across the decoded models (ClusterResource, GuestStatus, NodeStatus.Memory, BackupFile, StorageVolume, PhysicalDisk, ZFSPool, StorageSummary, ProxmoxTaskInfo, RRDPoint, Snapshot) to Int64. Small flags/counts/percentages stay Int. Add an Int64 overload for the byte formatters and guard the MB->bytes multiply in guest config against 32-bit overflow.

+++ App/Views/Guests/GuestDetailView.swift
@@ -268,7 +268,8 @@ struct GuestDetailView: View {
                 detailRow("Cores", cores + (config.sockets.map { " × \($0) sockets" } ?? ""))
             }
             if let memory = config.memoryMB {
-                detailRow("Memory", Format.bytes(memory * 1_048_576))
+                // Int64 so the MB->bytes multiply can't overflow 32-bit Int on watchOS.
+                detailRow("Memory", Format.bytes(Int64(memory) * 1_048_576))
             }
             if let os = config.osType { detailRow("OS type", os) }
             if let boot = config.bootOrder { detailRow("Boot order", boot) }
